+++ backend/src/main/java/com/usuario/sistemaBD/infrastructure/jpa/repository/UsuarioRepository.java
@@ -12,9 +12,9 @@ import java.util.Optional;
 public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
     Optional<Usuario> findByEmail(String email);
 
-    Long countByTipo(String tipo);
+    int countByTipo(String tipo);
 
-    Long countByTipoAndStatus(String tipo, String status);
+    int countByTipoAndStatus(String tipo, String status);
 
     @Transactional
     void deleteByEmail(String email);
